+++ commons-lang/src/test/java/com/hb0730/commons/lang/runtime/SystemUtilsTest.java
@@ -61,4 +61,10 @@ public class SystemUtilsTest {
             }
         });
     }
+
+    @Test
+    public void getPathSeparatorTest() {
+        String separator = SystemUtils.getPathSeparator();
+        Assert.assertNotNull(separator);
+    }
 }